Release 1.24 - 24/04/2025

+++ src/Web App/HE.Remediation.Core/UseCase/Areas/FireRiskAppraisal/ReportDetails/GetReportDetails/FRAEWReportDetails.cs
@@ -9,7 +9,7 @@ public class FRAEWReportDetails
     public decimal? FraewCost { get; set; }
     public string UndertakingFirm { get; set; }
     public int? NumberOfStoreys { get; set; }
-    public int? BuildingHeight { get; set; }
+    public decimal? BuildingHeight { get; set; }
     public ENoYes? BuildingInterimMeasures { get; set; }
     public EBasicComplexType? BasicComplexId { get; set; }
 }
